Refatoração do crud e criação da exceptionHandler

+++ src/main/java/br/com/central/infractions/service/RecordsService.java
@@ -1,8 +1,8 @@
 package br.com.central.infractions.service;
 
 import br.com.central.infractions.dto.RecordsDto;
-
 import br.com.central.infractions.entity.RecordsEntity;
+import br.com.central.infractions.exceptions.BusinessException;
 import br.com.central.infractions.repository.RecordsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -11,12 +11,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.NoResultException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import java.util.Optional;
 
 @Service
 public class RecordsService {
+
+    private final static Logger logger = Logger.getLogger(RecordsService.class.getName());
 
     @Autowired
     private RecordsRepository recordsRepository;
@@ -36,23 +38,37 @@ public class RecordsService {
 
     }
 
-    public RecordsEntity serchRecordsById(Long id) throws NoResultException{
+    public RecordsEntity serchRecordsById(Long id)throws BusinessException{
         Optional<RecordsEntity> records = recordsRepository.findById(id);
         if(records.isPresent()){
             return records.get();
         }
-        throw new NoResultException();
+        logger.log(Level.SEVERE,"register not found",BusinessException.class);
+        throw new BusinessException("register not found");
     }
 
     @Transactional
-    public void updateRecord(Long id, RecordsDto recordsDto){
+    public void updateRecord(Long id, RecordsDto recordsDto)throws BusinessException {
         Optional<RecordsEntity> findRecord = recordsRepository.findById(id);
-        RecordsEntity recordsEntity = findRecord.get();
-        recordsEntity.SetEntity(recordsDto);
-        recordsRepository.save(recordsEntity);
+        if (findRecord.isPresent()){
+            RecordsEntity recordsEntity = findRecord.get();
+            recordsEntity.SetEntity(recordsDto);
+            recordsRepository.save(recordsEntity);
+        }else {
+            logger.log(Level.SEVERE,"Unable to perform update",BusinessException.class);
+            throw new BusinessException("Unable to perform update");
+        }
+
     }
 
-    public void deleteRecords(Long id){
-        recordsRepository.deleteById(id);
+    public void deleteRecords(Long id)throws BusinessException{
+        Optional<RecordsEntity> findRecord = recordsRepository.findById(id);
+        if (findRecord.isPresent()){
+            recordsRepository.deleteById(id);
+        }else {
+            logger.log(Level.SEVERE,"Unable to delete record",BusinessException.class);
+            throw new BusinessException("Unable to delete record");
+        }
+
     }
 }
